Moving rotation structure and tests: misc -> preprocessing
